Added ivory coast zip files

+++ Dataset/Unzip_DS.py
@@ -1,5 +1,6 @@
 from zipfile import ZipFile
 import os 
+import warnings
 
 def UnzipFolders(domain):
     """
@@ -9,10 +10,13 @@ def UnzipFolders(domain):
             - domain: string with the name of the domain. Options: "Tanzania" or "IvoryCoast"
     """
     if (domain != "Tanzania") & (domain != "IvoryCoast"):
-        raise Exception("domain needs to be Tanzania or IvoryCoast")
+        raise Exception("Domain needs to be Tanzania or IvoryCoast")
 
     if len([i for i in os.listdir('.') if '.zip' in i]) != 0: 
-        for i in range(1): # NEEEDS TO BE CHANGED
-            with ZipFile(domain + str(i+1) + ".zip", 'r') as zipped:
-                zipped.extractall(path="./")
-            os.remove(domain + str(i+1) + ".zip")
+        for i in range(3): 
+            if (len([f for f in os.listdir('.') if domain + str(i+1) + ".zip" in f]) != 0):
+                with ZipFile(domain + str(i+1) + ".zip", 'r') as zipped:
+                    zipped.extractall(path="./")
+                os.remove(domain + str(i+1) + ".zip")
+            else:
+                raise warnings.warn("No zipped file found (" + domain + str(i+1) + ".zip" + ")")
